add command bar default function
